requirements.txt
# -----------------------------------------------------------------------------
# written by: Lawrence McDaniel
#             https://lawrencemcdaniel.com
#
# usage:      Python requirements for AWS Lambda functions. Create a virtual
#             environment in the root of this repository named `venv`. Terraform
#             modules will look for and include these requirements in the zip
#             packages for each Python-based Lambda function.
# -----------------------------------------------------------------------------

# dev and test
# ------------
boto3==1.34.34
botocore==1.34.34
pytest==8.0.0
pytest_mock==3.12.0
requests==2.31.0

# Code linters, formatters, and security scanners
# ------------
black==24.1.1
flake8==7.0.0
flake8-coding==1.3.2
pre-commit==3.6.0
isort==5.13.2
mypy==1.8.0
pylint==3.0.3
bandit==1.7.7
pydocstringformatter==0.7.3
tox==4.12.1
codespell==2.2.6
coverage==7.4.0

# Django
# ------------
Django==5.0.2
djangorestframework==3.14.0
markdown==3.5.2
django-filter==23.5

# Lambda layer: openai
# ------------
openai==1.11.1

# Lambda layer: common
# ------------
python-dotenv==1.0.1
pydantic==2.6.0
pydantic-settings==2.1.0
python-hcl2==4.3.2

# Lambda layer: langchain
# ------------
langchain==0.1.1
langchain-openai==0.0.5

# Lambda layer: nlp
# ------------
python-Levenshtein==0.24.0
pyyaml

# weather function
googlemaps==4.10.0
openmeteo-requests==1.1.0
requests-cache==1.1.1
retry-requests==2.0.0
pandas==2.2.0
